requirements.txt
mediapipe==0.10.9
numpy==1.24.3
opencv_contrib_python==4.8.0.74
Pillow==10.0.0
Pillow==10.2.0
streamlit==1.29.0
